More updates to the README
Added more elaboration on the IPAM controller.

The README is not now consistent with the code.  The README says a
NetworkAttachment's Spec includes a VNI, but this is not true.

We have not really made up our minds whether there is an implicit
router.  If there is then selection should be on VNI; if there is not
then selection should be on Subnet.  If would be possible to get the
effect of an Informer for a VNI from a collection of Informers for the
VNI's Subnets, but that would be icky.