Add spec.VNI and status.HostIP to NetworkAttachment
Add spec.VNI and status.HostIP to the fields of Custom Resource NetworkAttachment. The contents of pkc/client are up to date with the additions, meaning that hack/update-codegen.sh has been run. At this commit though, the API server extension has not been updated yet: the new fields introduce consitency requirements (a NetworkAttachment spec.VNI must be equal to spec.VNI of its Subnet) that the API server does not enforce yet. Also, the fields in NetworkAttachmentSpec and SubnetSpec are no longer optional.