Make apiserver extension's replicaset always pull image
In the replicaset controlling the API server extension the container
image pull policy for the said API server is set to "Always".

The reason is that the image with tag "latest" is used. That image
will be updated frequently, using a different pull policy means
missing the updates when re-deploying on the same cluster the API
extension server.