Fix bug making it impossible to resize subnets CIDR
During subnets validation, both for creates and updates a subnet is
compared against all the other subnets to make sure their CIDRs do not
overlap. Before this commit a subnet was compared against itself as
well, thus making it impossible to update it by resizing its CIDR to
a subset of its previous addresses pool. This commit fixes that bug.