ConnAgent: re-design/re-implementation of logic. Make logger thread-safe
Re-implement the connection agent logic according to a different design.
This new version of the connection agent makes the following guarantees:
(1) for a network attachment, at most one interface is implemented within
a single node at every time, (2) in case of updates to the vni field and/or
updates to the node field which trigger a change in remoteness, if the updates
stop at some point the interface for the stable version of the attachment
is created. In the latter case however, there's no guarantee as for the
order in which "transient interfaces" are created: it might be different
wrt the order of the updates to the attachment.

This commit also makes the logger network fabric thread-safe.