ConnAgent: re-implement logic with new design and small fixes
Re-implement the connection agent logic with a redically different design.
References are namespaced name and carry no vni or locality flag. Notification
handlers for remote attachments add/remove the vni with which they see the
attachment to a map from namespaced name to vnis where that namespaced name
has been seen. Workers processing namespaced name look at the vnis where the
namespaced name has been seen. If more than one vni is found, the worker stops
processing because the current attachment version is ambiguous, and the
namespaced name is requeued with a delay. The hope is that by the time the
namespaced name is processed again the ambiguity has been solved. This commit
also contains some small fixes to the connection agent.