OvS fabric: TCP instead of UDS
The OvS CLI utilities can choose to either use TCP sockets or Unix Domain
Sockets to communicate with the openvswitch database server and to install
OpenFlow flows. This commit switches from UDSs to TCP. The reason is that
the openvswitch database server runs on the host, not in the Connection
Agent container. Hence, using UDSs requires mounting the directory where
all the database server creates its UDSs. If the openvswitch crashes on the
host, such directory is deleted and recreated, breaking the directory mount
in the connection agent container. This makes communication with openvswitch
impossible for the connection agent. Switching to TCP sockets solves the issue.
The connection agent Pod runs in the same network namespace as the host. The
openvswitch database server can bind to a TCP port P. The connection agent can
contact the database server at localhost:P. This makes sharing volumes unneeded,
and the connection agent robust to crashes and restarts of Openvswith. The same
considerations apply for installing OpenFlow flows.